refactor(codegen): split field-init helpers out of new.rs to clear 2000-LOC gate
The per-new-site allocator outlining pushed lower_call/new.rs to 2038 lines,
tripping the check_file_size.sh CI gate. Move FieldInitMode + the
apply_field_initializers_recursive walker (pure move, no behavior change)
into a sibling field_init.rs and re-export from mod.rs; new.rs drops to 1736.
